+++ files/XmlFileReadWriter.ts
@@ -1,11 +1,11 @@
 ﻿import DomBuilderHelper = require("@twg2/dom-builder/dom/DomBuilderHelper");
-import XmlFileInst = require("./XmlFileInst");
+import { XmlFileInst } from "./XmlFileInst";
 
 /** An OpenXmlIo FileReadWriter implementation with a configurable pre-write callback and a cache containing the last read()/loadFromDom() result
  * @author TeamworkGuy2
  * @since 2016-5-27
  */
-class XmlFileReadWriter<T> implements OpenXmlIo.FileReadWriter<T> {
+export class XmlFileReadWriter<T> implements OpenXmlIo.FileReadWriter<T> {
     public fileInfo: OpenXmlIo.XlsxFileType;
     private prepForWrite: (xmlDoc: OpenXmlIo.WriterContext, inst: T) => void;
     private rootReadWriter: OpenXmlIo.ReadWrite<T>;
@@ -97,5 +97,3 @@ class XmlFileReadWriter<T> implements OpenXmlIo.FileReadWriter<T> {
     }
 
 }
-
-export = XmlFileReadWriter;